requirements.txt
dash
dash-bootstrap-components
pandas
numpy
yfinance
plotly
kaleido
reportlab
python-dotenv
firebase-admin
Flask-Session
requests